+++ virtual-keyboard/src/js/get-template.js
